Modularize NavBar
